# app/api/chat/route.ts
import { NextRequest, NextResponse } from "next/server";
import Groq from "groq-sdk";
import { sendChatMessageToApi } from "@/lib/services/api/chat";
import { ApiChatMessage } from "@/lib/types/api/chat";

// ISPM ORIENT'IA System Prompt (used for local fallback or system context)
const SYSTEM_PROMPT = `Tu es ORIENT'IA, l'assistant virtuel intelligent d'orientation pédagogique de l'ISPM (Institut Supérieur Polytechnique de Madagascar).

Ton rôle exclusif est d'aider les candidats à choisir la meilleure filière de formation parmi les 16 formations officielles de l'ISPM :

**Mention : Informatique et Télécommunications**
- IGGLIA : Informatique de Gestion, Génie Logiciel et Intelligence Artificielle
- ESIIA : Électronique, Systèmes Informatiques et Intelligence Artificielle
- IMTICIA : Informatique Multimédia, TIC et Intelligence Artificielle
- ISAIA : Informatique Statistique Appliquée et Intelligence Artificielle

**Mention : Génie Industriel**
- EMII : Électromécanique et Informatique Industrielle
- ICMP : Industries Chimiques, Minières et Pétrolières

**Mention : Génie Civil et Architecture**
- GCA : Génie Civil et Architecture

**Mention : Droit et Techniques des Affaires**
- CAA : Commerce et Administration des Affaires
- EMP : Économie et Management de Projet
- FIC : Finances et Comptabilité
- DTJA : Droit et Techniques Juridiques des Affaires

**Mention : Biotechnologie et Agronomie**
- IAA : Industries Agroalimentaires
- AEE : Agriculture et Élevage
- PIP : Pharmacologie et Industries Pharmaceutiques

**Mention : Tourisme**
- TEH : Tourisme, Environnement et Hôtellerie
- TEE : Tourisme et Environnement

---

## Format de tes réponses :
- Réponses structurées, complètes et détaillées sans coupure.
- Utilise Markdown (titres ###, listes à puces, texte en gras **...**).
- En fin de recommandation, inclus toujours la mention légale d'orientation.`;

export async function POST(req: NextRequest) {
  try {
    const { messages, profile } = await req.json();

    if (!messages || !Array.isArray(messages) || messages.length === 0) {
      return NextResponse.json({ error: "Aucun message fourni." }, { status: 400 });
    }

    // Extract last user message and formatted conversation history
    const userMessages = messages.filter((m: { sender: string; content: string }) => m.content?.trim());
    const lastUserMessage = userMessages[userMessages.length - 1];
    const messageText = lastUserMessage?.content || "";

    const conversationHistory: ApiChatMessage[] = userMessages
      .slice(0, -1)
      .slice(-4) // keep last 4 messages to stay light and avoid Groq context timeouts
      .map((m: { sender: string; content: string }) => ({
        role: m.sender === "user" ? ("user" as const) : ("assistant" as const),
        content: m.content,
      }));

    // Format profile details directly into the message text so any version of backend API reads it
    let messageWithProfile = messageText;
    if (profile) {
      const grades = Array.isArray(profile.academicGrades) && profile.academicGrades.length > 0
        ? profile.academicGrades.map((g: { subject: string; grade: number }) => `${g.subject}: ${g.grade}/20`).join(", ")
        : "Non spécifiées";
      const subjects = Array.isArray(profile.preferredSubjects) && profile.preferredSubjects.length > 0
        ? profile.preferredSubjects.join(", ")
        : "Non spécifiées";
      const skills = Array.isArray(profile.declaredSkills) && profile.declaredSkills.length > 0
        ? profile.declaredSkills.join(", ")
        : "Non spécifiées";
      const interests = Array.isArray(profile.interests) && profile.interests.length > 0
        ? profile.interests.join(", ")
        : "Non spécifiés";

      messageWithProfile = `[PROFIL CANDIDAT RÉEL SAISI SUR L'APPLICATION] :
- Nom : ${profile.name || "Candidat"}
- Niveau d'études / Bac : ${profile.currentLevel || "Non spécifié"}
- Notes académiques obtenues : ${grades}
- Matières préférées : ${subjects}
- Compétences déclarées : ${skills}
- Centres d'intérêt : ${interests}
- Environnement / Domaine visé : ${profile.preferredWorkEnvironment || "Non spécifié"}

[QUESTION DU CANDIDAT] :
${messageText}`;
    }

    // Build exact profil_candidat expected by backend API schema
    // serie_bac must be a short value like "C", "D", "S", "Technique", not a full sentence
    let profilCandidat = undefined;
    if (profile) {
      const grades = Array.isArray(profile.academicGrades) && profile.academicGrades.length > 0
        ? profile.academicGrades
        : [];

      const avgMoyenne = grades.length > 0
        ? Math.round((grades.reduce((acc: number, curr: { grade: number }) => acc + curr.grade, 0) / grades.length) * 10) / 10
        : 14.0;

      const matieresFortes = Array.isArray(profile.preferredSubjects) && profile.preferredSubjects.length > 0
        ? profile.preferredSubjects.slice(0, 5) // limit to 5 max
        : grades.filter((g: { grade: number }) => g.grade >= 14).map((g: { subject: string }) => g.subject).slice(0, 5);

      // Extract short serie_bac code from currentLevel string
      // currentLevel examples: "Baccalauréat — Série C", "Licence 3 (ISPM) — Parcours IGGLIA"
      let serieBac = "C"; // safe default
      const lvl = profile.currentLevel || "";
      const serieMatch = lvl.match(/Série\s+([A-Z0-9]+(?:\s+[A-Z][A-Za-zÀ-ú]*)*)/);
      if (serieMatch) {
        // e.g. "Série C" → "C", "Série OSE" → "OSE", "Série Technique" → "Technique"
        serieBac = serieMatch[1].trim();
      } else if (lvl.toLowerCase().includes("licence") || lvl.toLowerCase().includes("master")) {
        // For Bac+3 and Master1 profiles, the serie doesn't apply — send "L3" or "M1"
        if (lvl.toLowerCase().includes("master")) serieBac = "M1";
        else serieBac = "L3";
      }

      profilCandidat = {
        serie_bac: serieBac,
        moyenne: avgMoyenne,
        matieres_fortes: matieresFortes.length > 0 ? matieresFortes : ["Mathématiques"],
      };
    }


    // ─── STEP 1: Attempt real FastAPI backend call ─────────────────────────────
    // Send pure messageText (exactly like Swagger UI) because backend main.py
    // injects profil_candidat into the prompt and queries ChromaDB with message!
    const apiPayload = {
      message: messageText,
      conversation_history: conversationHistory,
      profil_candidat: profilCandidat,
      top_k: 20,
    };
    console.log("[ORIENT'IA] Payload envoyé à Render (identique Swagger):", JSON.stringify(apiPayload, null, 2));

    try {
      const fastApiResponse = await sendChatMessageToApi(apiPayload);

      if (fastApiResponse && fastApiResponse.answer) {
        return NextResponse.json({
          content: fastApiResponse.answer,
          sources: fastApiResponse.sources || [],
          requestId: fastApiResponse.request_id,
          disclaimer: fastApiResponse.disclaimer,
          backendSource: "fastapi_render",
        });
      }
    } catch (fastApiErr: any) {
      // Log full error detail including Render 500 response body
      const detail = fastApiErr?.details ? JSON.stringify(fastApiErr.details) : fastApiErr?.message || String(fastApiErr);
      console.error("[ORIENT'IA] Erreur FastAPI Render:", fastApiErr?.status, detail);
      console.warn("[ORIENT'IA] Basculement sur moteur Groq local.");
    }

    // ─── STEP 2: Fallback to local Groq + System Prompt ────────────────────────
    if (!process.env.GROQ_API_KEY) {
      return NextResponse.json(
        { error: "Clé API Groq et service FastAPI indisponibles." },
        { status: 500 }
      );
    }

    const groq = new Groq({ apiKey: process.env.GROQ_API_KEY });

    const groqMessages = [
      { role: "system" as const, content: SYSTEM_PROMPT },
      ...messages
        .filter((m: { sender: string; content: string }) => m.content?.trim())
        .map((m: { sender: string; content: string }) => ({
          role: m.sender === "user" ? ("user" as const) : ("assistant" as const),
          content: m.content,
        })),
    ];

    const completion = await groq.chat.completions.create({
      model: "openai/gpt-oss-120b",
      messages: groqMessages,
      temperature: 0.4,
      max_tokens: 4096,
    });

    const content =
      completion.choices[0]?.message?.content ||
      "Je suis désolé, je n n'ai pas pu générer une réponse. Veuillez réessayer.";

    return NextResponse.json({
      content,
      backendSource: "groq_local_fallback",
    });
  } catch (error: unknown) {
    const errorMsg = error instanceof Error ? error.message : "Erreur interne.";
    return NextResponse.json({ error: errorMsg }, { status: 500 });
  }
}