# openapi.json
{"openapi":"3.1.0","info":{"title":"ORIENT’IA","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/chat":{"post":{"summary":"Chat","operationId":"chat_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ChatRequest":{"properties":{"message":{"type":"string","title":"Message"},"conversation_history":{"items":{"$ref":"#/components/schemas/Message"},"type":"array","title":"Conversation History","default":[]},"top_k":{"type":"integer","title":"Top K","default":3}},"type":"object","required":["message"],"title":"ChatRequest"},"ChatResponse":{"properties":{"answer":{"type":"string","title":"Answer"},"request_id":{"type":"string","title":"Request Id"},"sources":{"items":{"$ref":"#/components/schemas/SourceMetadata"},"type":"array","title":"Sources","default":[]},"disclaimer":{"type":"string","title":"Disclaimer","default":"ORIENT’IA est un outil d’aide à l’orientation. Ses recommandations ne remplacent ni l’avis d’un conseiller pédagogique ni une décision officielle d’admission."}},"type":"object","required":["answer","request_id"],"title":"ChatResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Message":{"properties":{"role":{"type":"string","title":"Role"},"content":{"type":"string","title":"Content"}},"type":"object","required":["role","content"],"title":"Message"},"SourceMetadata":{"properties":{"code_parcours":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Code Parcours"},"nom_parcours":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Nom Parcours"},"mention":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mention"},"fichier_source":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Fichier Source"},"source_titre":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source Titre"},"source_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source Url"},"statut":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Statut"},"score":{"type":"number","title":"Score"}},"type":"object","required":["score"],"title":"SourceMetadata"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}